Bayesian Project
The purpose of this analysis will be to determine if cardiovascular disease is a more or less prevalent cause of death in the United States than elsewhere, that is if the proportion of deaths due to cardiovascular disease is greater or smaller in the US than in other countries. We are also interested in comparing death due to cardiovascular disease in the US vs China, the countries of origin of the authors. According to the data character, a beta-binomial Bayesian model is chosen.  
The conclusion is that the proportion of deaths in the United States due to cardiovascular disease is lower than the same cause in China and all other countries as a whole. There is further research to be done in order to determine the reason of our conclusion.